test(agent): classify replay intrinsic boundary
